b=13979 i=Adilger
change tdir to start from d0 to be cleaned up automatically

+++ lustre/tests/test-framework.sh
@@ -1002,8 +1002,7 @@ cleanup_and_setup_lustre() {
 
 check_and_cleanup_lustre() {
     if [ "`mount | grep $MOUNT`" ]; then
-        rm -rf $DIR/[Rdfs][0-9]*
-        rm -f $DIR/${TESTSUITE}/[Rdfs][1-9]*
+        [ -n "$DIR" ] && rm -rf $DIR/[Rdfs][0-9]*
     fi
     if [ "$I_MOUNTED" = "yes" ]; then
         cleanupall -f || error "cleanup failed"
@@ -1328,7 +1327,7 @@ run_one() {
     testnum=$1
     message=$2
     tfile=f${testnum}
-    export tdir=d${TESTSUITE}/d${base}
+    export tdir=d0.${TESTSUITE}/d${base}
     local SAVE_UMASK=`umask`
     umask 0022
     mkdir -p $DIR/$tdir
